Fixed a few comments and excpetion messages. Added marker interface for Tnccs and Im data to restrict generics to them (e.g. read/writer).

+++ message/src/main/java/org/ietf/nea/pb/batch/PbBatchHeader.java
@@ -7,13 +7,13 @@ import de.hsbremen.tc.tnc.tnccs.batch.TnccsBatchHeader;
 
 public class PbBatchHeader implements TnccsBatchHeader{
 
-	 	private final byte version;                       //  8 bit(s)
+	 	private final short version;                       //  8 bit(s)
 	    private final PbBatchDirectionalityEnum directionality;  //  1 bit(s)
 	    private final PbBatchTypeEnum type;                      //  4 bit(s)
 	    
 	    private final long length;
 
-	    PbBatchHeader(final byte version, final PbBatchDirectionalityEnum directionality,
+	    PbBatchHeader(final short version, final PbBatchDirectionalityEnum directionality,
 				final PbBatchTypeEnum type, final long length) {
 			this.version = version;
 	    	this.directionality = directionality;
