+++ message/src/main/java/org/ietf/nea/pb/message/PbMessageValueAccessRecommendationBuilder.java
@@ -6,7 +6,7 @@ import de.hsbremen.tc.tnc.tnccs.message.TnccsMessageValueBuilder;
 
 public interface PbMessageValueAccessRecommendationBuilder extends TnccsMessageValueBuilder{
 
-	public abstract PbMessageValueAccessRecommendationBuilder setRecommendation(short recommendation)
+	public abstract PbMessageValueAccessRecommendationBuilder setRecommendation(int recommendation)
 			throws RuleException;
 
 }
